+++ game-jurassicworldevolution3/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None Planned
 
+## [0.2.0] - 2026-06-06
+
+- Added: Xbox version support
+- Fixed: ACSE Mods always installed with a top level folder
+
 ## [0.1.0] - 2025-XX-XX
 
 - Initial release